* commit '42f64ac55735cec60882d24dfff6e45eb95770fa':
  fix for Tesnet: deposited to many coints. Increasing gas pice to 2 gwei.

# Conflicts:
#	app/src/main/java/com/teambrella/android/util/TeambrellaUtilService.java

+++ app/src/main/java/com/teambrella/android/blockchain/EtherAPI.java
@@ -2,6 +2,8 @@ package com.teambrella.android.blockchain;
 
 import com.google.gson.JsonObject;
 
+import java.math.BigInteger;
+
 import retrofit2.Call;
 import retrofit2.http.Field;
 import retrofit2.http.FormUrlEncoded;
@@ -23,5 +25,5 @@ public interface EtherAPI {
 
 
     @GET("api?module=account&action=balance")
-    Call<Scan<Long>> checkBalance(@Query("address") String address);
+    Call<Scan<BigInteger>> checkBalance(@Query("address") String address);
 }
